-Warning fixes at suggestion of tezeriusz -Added small always-active mining bonus to homeworld special -Minor code unduplication in SidePanel.cpp
svn path=/trunk/FreeOrion/; revision=3616

+++ UI/CombatSetupWnd.cpp
@@ -108,7 +108,7 @@ namespace {
             if (map_it != effect_accounting_map.end())
                 meter_map = &(map_it->second);
 
-            for (int i = 0; i < meters.size(); ++i) {
+            for (size_t i = 0; i < meters.size(); ++i) {
                 const MeterType METER_TYPE = meters[i];
                 const MeterType ASSOCIATED_METER_TYPE = associated_meters[i];
